Reporting of internal free space is wrong
In my testing, line 72 does not correctly report the free space remaining on the internal storage partition with the hardcoded path of "/". After reading through a response on Stack Overflow (http://stackoverflow.com/questions/4799643/getting-all-the-total-and-available-space-on-android) I tried changing this line to get the root path using the Environment object and I'm now able to access persistent storage with the benefit of checking available space first as requestFileSystem was intended to.